Moved common functions to new module and out of app.ts
