+++ server/app.ts
@@ -1,27 +1,14 @@
 import * as fs from "fs";
 import * as path from "path";
-import * as url from "url";
-import * as os from "os";
 import * as crypto from "crypto";
-import * as http from "http";
 
 import * as express from "express";
 import * as serveStatic from "serve-static";
 import * as compression from "compression";
-import * as bodyParser from "body-parser";
 import * as cookieParser from "cookie-parser";
-import * as multer from "multer";
 import * as morgan from "morgan";
-import * as ajv from "ajv";
 
-const PORT = parseInt(process.env.PORT) || 3000;
-const MONGO_URL = process.env.MONGO_URL || "mongodb://localhost/";
-const UNIQUE_APP_ID = process.env.UNIQUE_APP_ID || "registration";
-
-export const STATIC_ROOT = path.resolve(__dirname, "../client");
-export const VERSION_NUMBER = JSON.parse(fs.readFileSync(path.resolve(__dirname, "../package.json"), "utf8")).version;
-export const VERSION_HASH = require("git-rev-sync").short();
-
+// Set up Express and its middleware
 export let app = express();
 app.use(morgan("dev"));
 app.use(compression());
@@ -33,11 +20,12 @@ let cookieParserInstance = cookieParser(undefined, {
 });
 app.use(cookieParserInstance);
 
-import * as mongoose from "mongoose";
-(<any>mongoose).Promise = global.Promise;
-mongoose.connect(url.resolve(MONGO_URL, UNIQUE_APP_ID));
-export {mongoose}; // For future unit testing and dependent routes; see https://github.com/HackGT/Ultimate-Checkin/blob/master/test/api.ts#L11
-
+import {
+	// Functions
+	pbkdf2Async,
+	// Constants
+	PORT, STATIC_ROOT, VERSION_NUMBER, VERSION_HASH
+} from "./common";
 import {
 	IUser, IUserMongoose, User
 } from "./schema";
@@ -71,101 +59,6 @@ Created default admin user
 **Delete this user after you have used it to set up your account**
 	`);
 })();
-
-// Promise version of crypto.pbkdf2()
-export function pbkdf2Async (...params: any[]) {
-	return new Promise<Buffer>((resolve, reject) => {
-		params.push(function (err: Error, derivedKey: Buffer) {
-			if (err) {
-				reject(err);
-				return;
-			}
-			resolve(derivedKey);
-		});
-		crypto.pbkdf2.apply(null, params);
-	});
-}
-
-export let postParser = bodyParser.urlencoded({
-	extended: false
-});
-export let uploadHandler = multer({
-	"storage": multer.diskStorage({
-		destination: function (req, file, cb) {
-			// The OS's default temporary directory
-			// Should be changed (or moved via fs.rename()) if the files are to be persisted
-			cb(null!, os.tmpdir());
-		},
-		filename: function (req, file, cb) {
-			cb(null!, `${file.fieldname}-${Date.now()}.${path.extname(file.originalname)}`);
-		}
-	}),
-	"limits": {
-		"fileSize": 50000000, // 50 MB
-		"files": 1
-	}
-});
-
-// For API endpoints
-export let authenticateWithReject = async function (request: express.Request, response: express.Response, next: express.NextFunction) {
-	let authKey = request.cookies.auth;
-	let user = await User.findOne({"auth_keys": authKey});
-	if (!user) {
-		response.status(401).json({
-			"error": "You must log in to access this endpoint"
-		});
-	}
-	else {
-		response.locals.email = user.email;
-		next();
-	}
-};
-// For directly user facing endpoints
-export let authenticateWithRedirect = async function (request: express.Request, response: express.Response, next: express.NextFunction) {
-	let authKey = request.cookies.auth;
-	let user = await User.findOne({"auth_keys": authKey});
-	if (!user) {
-		response.redirect("/login");
-	}
-	else {
-		response.locals.email = user.email;
-		next();
-	}
-};
-
-export function readFileAsync (filename: string): Promise<string> {
-	return new Promise<string>((resolve, reject) => {
-		fs.readFile(filename, "utf8", (err, data) => {
-			if (err) {
-				reject(err);
-				return;
-			}
-			resolve(data);
-		})
-	});
-}
-
-// Validates and returns JSON object of questions
-export let validateQuestions = async function (questionsFile: string, schemaFile: string): Promise<any> {
-	let questions, schema;
-	try {
-		questions = JSON.parse(await readFileAsync(questionsFile));
-		schema = JSON.parse(await readFileAsync(schemaFile));
-	}
-	catch (err) {
-		// Invalid JSON or file read failed unexpectedly
-		return Promise.reject(err);
-	}
-
-	let validator = new ajv();
-	let valid = validator.validate(schema, questions);
-	if (!valid) {
-		return Promise.reject(validator.errors);
-	}
-	else {
-		return questions;
-	}
-};
 
 let apiRouter = express.Router();
 // API routes go here
